Metodo Criar Nova Conta
Quando criado uma nova conta, será criado um movimento;

Quando a nova conta for transferência de outra conta, alterar saldo da conta de transferência criando outro movimento;

Inserido forma de não deletar as Classes - Nova Conta e Transferência nem Conta - Dinheiro.

+++ Controller/NConta_DAO.cs
@@ -37,7 +37,7 @@ namespace Controle_Financeiro_Pessoal.Controller
             // definindo sql de insersão e atribuindo os parâmetros
             DBContas.cmd = new SqlCommand("insert into Conta(Conta, Saldo) values (@conta, @saldo)", DBContas.sqlConn);
             DBContas.cmd.Parameters.Add("@conta", SqlDbType.VarChar).Value = conta;
-            DBContas.cmd.Parameters.Add("@saldo", SqlDbType.Int).Value = saldo;
+            DBContas.cmd.Parameters.Add("@saldo", SqlDbType.Decimal).Value = saldo;
 
             if (conexão(DBContas.cmd))
                 return true;
